pyproject.toml
[project]
name = "medseg-api"
version = "0.1.0"
description = "Production medical image segmentation API serving hybrid CNN-ViT lung segmentation models"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "timm>=1.0.0",
    "opencv-python-headless>=4.8.0",
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.9",
    "mlflow>=2.10.0",
    "pyyaml>=6.0",
    "httpx>=0.27.0",
    "albumentations>=1.3.0",
    "pydicom>=2.4.0",
    "pandas>=2.0.0",
    "plotly>=6.6.0",
    "statsmodels>=0.14.0",
    "streamlit>=1.55.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
    "httpx>=0.27.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
